pyproject.toml
[tool.poetry]
name = "projectx"
version = "0.1.0"
description = ""
authors = ["IvanVinitskyi <ivan.vinitskyi@uprisehealth.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.6"
django-ninja = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
